Refactor early hackery into a more usable form
