+++ bracket-bevy/src/builder/terminal_layer.rs
@@ -1,0 +1,4 @@
+#[derive(Clone)]
+pub enum TerminalLayer {
+    Simple { font_index: usize, width: usize, height: usize },
+}
