+++ bracket-bevy/src/consoles/mod.rs
@@ -1,0 +1,4 @@
+mod simple_console;
+pub(crate) use simple_console::*;
+mod update_system;
+pub(crate) use update_system::*;
