+++ bracket-bevy/src/fonts/mod.rs
@@ -1,0 +1,3 @@
+mod font_store;
+pub(crate) use font_store::*;
+
